+++ src/main.js
@@ -11,9 +11,9 @@ import FilmDetails from './view/film-details.js';
 import { generateFilmCard } from './mock/film-info.js';
 import { generateComments } from './mock/comments.js';
 import Comments from './view/comment.js';
-import { render, RenderPosition } from './utils.js';
 import StatisticSection from './view/statistic-section.js';
 import EmptyFilmSection from './view/empty-film-list.js';
+import { remove, render, RenderPosition } from './utils/render.js';
 
 const FILMS_NUMBER = 17;
 const FILMS_IN_LINE = 5;
@@ -31,7 +31,7 @@ const comments = new Array(COMMENTS_NUMBER).fill().map(generateComments);
 const filmsSectionComponent = new FilmsSection();
 
 // Render profile info
-render(headerElement, new ProfileInfo().getElement(), RenderPosition.BEFOREEND);
+render(headerElement, new ProfileInfo(), RenderPosition.BEFOREEND);
 
 // Render main navigation
 const countFilters = () => {
@@ -48,20 +48,20 @@ const countFilters = () => {
   return counter;
 };
 
-render(mainElement, new MainNavigation(countFilters()).getElement(), RenderPosition.BEFOREEND);
+render(mainElement, new MainNavigation(countFilters()), RenderPosition.BEFOREEND);
 
 // Render films section
 if (filmCards.length === 0) {
-  render(mainElement, new EmptyFilmSection().getElement(), RenderPosition.BEFOREEND);
+  render(mainElement, new EmptyFilmSection(), RenderPosition.BEFOREEND);
 } else {
-  render(mainElement, filmsSectionComponent.getElement(), RenderPosition.BEFOREEND);
+  render(mainElement, filmsSectionComponent, RenderPosition.BEFOREEND);
   const filmsListContainer = filmsSectionComponent.getElement().querySelector('.films-list__container');
 
   const renderFilmCard = (filmList, card) => {
     const filmCardComponent = new FilmCard(card);
 
     // Render filmcard with listeners
-    render(filmList, filmCardComponent.getElement(), RenderPosition.BEFOREEND);
+    render(filmList, filmCardComponent, RenderPosition.BEFOREEND);
     filmCardComponent.setPopupOpenHandler(() => {
       siteBodyElement.appendChild(filmPopup.getElement());
       siteBodyElement.classList.toggle('hide-overflow');
@@ -70,9 +70,8 @@ if (filmCards.length === 0) {
 
       document.addEventListener('keydown', onEscKeyDown);
       filmPopup.setCloseBtnClickHandler(() => {
-        siteBodyElement.removeChild(filmPopup.getElement());
         siteBodyElement.classList.toggle('hide-overflow');
-        filmPopup.removeElement();
+        remove(filmPopup);
         document.removeEventListener('keydown', onEscKeyDown);
       });
     });
@@ -84,7 +83,7 @@ if (filmCards.length === 0) {
   const onEscKeyDown = (evt) => {
     if (evt.key === 'Escape' || evt.key === 'Esc') {
       evt.preventDefault();
-      siteBodyElement.removeChild(filmPopup.getElement());
+      remove(filmPopup);
       siteBodyElement.classList.toggle('hide-overflow');
       document.removeEventListener('keydown', onEscKeyDown);
     }
@@ -122,7 +121,7 @@ if (filmCards.length === 0) {
     let renderedCardCount = FILMS_IN_LINE;
     const showMoreBtn = new ShowMoreButton();
 
-    render(filmsSectionComponent.getElement().querySelector('.films-list'), showMoreBtn.getElement(), RenderPosition.BEFOREEND);
+    render(filmsSectionComponent.getElement().querySelector('.films-list'), showMoreBtn, RenderPosition.BEFOREEND);
 
     showMoreBtn.setClickHandler(() => {
       filmCards
@@ -132,7 +131,7 @@ if (filmCards.length === 0) {
       renderedCardCount += FILMS_IN_LINE;
 
       if (renderedCardCount >= filmCards.length) {
-        showMoreBtn.getElement().remove();
+        remove(showMoreBtn);
       }
     });
   }
@@ -140,9 +139,9 @@ if (filmCards.length === 0) {
 
 // Render statistic
 const statisticSection = new StatisticSection();
-render(mainElement, statisticSection.getElement(), RenderPosition.BEFOREEND);
-render(statisticSection.getElement(), new StatisticRank().getElement(), RenderPosition.BEFOREEND);
-render(statisticSection.getElement(), new StatisticFilter().getElement(), RenderPosition.BEFOREEND);
+render(mainElement, statisticSection, RenderPosition.BEFOREEND);
+render(statisticSection, new StatisticRank(), RenderPosition.BEFOREEND);
+render(statisticSection, new StatisticFilter(), RenderPosition.BEFOREEND);
 
 // Statistic counter
 const countStatistic = () => {
@@ -177,5 +176,5 @@ const countStatistic = () => {
   return counter;
 };
 
-render(statisticSection.getElement(), new StatisticText(countStatistic()).getElement(), RenderPosition.BEFOREEND);
-render(footerStatisticsElement, new FooterStats(FILMS_NUMBER).getElement(), RenderPosition.BEFOREEND);
+render(statisticSection, new StatisticText(countStatistic()), RenderPosition.BEFOREEND);
+render(footerStatisticsElement, new FooterStats(FILMS_NUMBER), RenderPosition.BEFOREEND);
