+++ src/view/film-card.js
@@ -1,5 +1,5 @@
 import dayjs from 'dayjs';
-import { checkPlural, formatingRuntime } from '../utils';
+import { checkPlural, formatingRuntime } from '../utils/common';
 import AbstractView from './abstract';
 
 const CARD_DESCRIPTION_LENGTH = 140;
